backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.5
numpy==2.1.2
pandas==2.2.3
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.6
